requirements.txt
numpy<2.0
onnx
onnxruntime
PySide6
pillow
tqdm
scikit-learn
matplotlib
torchinfo
jupyterlab
ipywidgets
